extract normal state component out of Start
