+++ qt/launcher-gui/src/routes/Start.svelte
@@ -3,79 +3,29 @@ Copyright: Ankitects Pty Ltd and contributors
 License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 -->
 <script lang="ts">
-    import {
-        Mirror,
-        type Options as OptionsProto,
-        type ChooseVersionResponse,
-        type GetLangsResponse_Pair,
-        type GetMirrorsResponse_Pair,
-    } from "@generated/anki/launcher_pb";
-    import {
-        chooseVersion,
-        getAvailableVersions,
-        getExistingVersions,
-    } from "@generated/backend-launcher";
+    import { type GetLangsResponse_Pair } from "@generated/anki/launcher_pb";
     import Row from "$lib/components/Row.svelte";
     import EnumSelectorRow from "$lib/components/EnumSelectorRow.svelte";
     import SettingTitle from "$lib/components/SettingTitle.svelte";
     import TitledContainer from "$lib/components/TitledContainer.svelte";
     import Container from "$lib/components/Container.svelte";
     import { tr } from "./stores";
-    import Warning from "./Warning.svelte";
-    import Action from "./Action.svelte";
-    import Spinner from "./Spinner.svelte";
-    import Options from "./Options.svelte";
-    import Term from "./Term.svelte";
-    import AnkiWillStart from "./AnkiWillStart.svelte";
-    import type { Terminal } from "@xterm/xterm";
 
     let {
         langs,
         selectedLang = $bindable(),
-        options,
-        mirrors,
+        children,
+        footer,
     }: {
         langs: GetLangsResponse_Pair[];
         selectedLang: string;
-        options: OptionsProto;
-        mirrors: GetMirrorsResponse_Pair[];
+        children: any;
+        footer: any;
     } = $props();
-
-    let releasesPromise = $state(getAvailableVersions({}, { alertOnError: false }));
-    let existingPromise = $state(getExistingVersions({}, { alertOnError: false }));
-    let loadPromise = $derived(Promise.all([releasesPromise, existingPromise]));
 
     const availableLangs = $derived(
         langs.map((p) => ({ label: p.name, value: p.locale })),
     );
-
-    let allowBetas = $state(options.allowBetas);
-    let downloadCaching = $state(options.downloadCaching);
-    let selectedMirror = $state(Mirror.DISABLED);
-
-    let choosePromise: Promise<ChooseVersionResponse | null> = $state(
-        Promise.resolve(null),
-    );
-
-    let error: Error | null = $state(null);
-    const setError = (e: Error) => {
-        error = e;
-    };
-
-    let term: Terminal | undefined = $state(undefined);
-    let termOpen = $state(false);
-    let chosen = $state(false);
-
-    const choose = (version: string, keepExisting: boolean, current?: string) => {
-        chosen = true;
-        term?.reset();
-        choosePromise = chooseVersion({
-            version,
-            keepExisting,
-            options: { allowBetas, downloadCaching, mirror: selectedMirror },
-            ...(current ? { current } : {}),
-        });
-    };
 </script>
 
 <Container
@@ -101,55 +51,12 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
                     {$tr.launcherLanguage()}
                 </SettingTitle>
             </EnumSelectorRow>
-            {#await choosePromise}
-                <Row class="centre m-3">
-                    <Spinner label={$tr.launcherSyncing()} />
-                </Row>
-            {:then res}
-                {#if res === null}
-                    {#await loadPromise}
-                        <Row class="centre m-3">
-                            <Spinner label={$tr.launcherLoadingVersions()} />
-                        </Row>
-                    {:then [releases, existing]}
-                        <Action {releases} {existing} {allowBetas} {choose} />
-                    {:catch e}
-                        {setError(e)}
-                        <Warning
-                            warning={$tr.lauuncherFailedToLoadVersions()}
-                            className="alert-danger"
-                        />
-                    {/await}
-                {:else}
-                    <Row class="centre m-3">
-                        <AnkiWillStart {res} />
-                    </Row>
-                {/if}
-            {:catch e}
-                {setError(e)}
-                <Warning
-                    warning={$tr.launcherFailedToSync()}
-                    className="alert-danger"
-                />
-            {/await}
-            {#if error != null}
-                <Row>
-                    <pre>{error.message}</pre>
-                </Row>
-            {/if}
+
+            {@render children?.()}
         </TitledContainer>
     </Row>
-    <Term bind:term bind:open={termOpen} />
-    {#if !chosen}
-        <Row class="row-columns">
-            <Options
-                {mirrors}
-                bind:allowBetas
-                bind:downloadCaching
-                bind:selectedMirror
-            />
-        </Row>
-    {/if}
+
+    {@render footer?.()}
 </Container>
 
 <style lang="scss">
@@ -181,13 +88,5 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
     :global(.centre) {
         justify-content: center;
-    }
-
-    pre {
-        white-space: pre-wrap; /* Since CSS 2.1 */
-        white-space: -moz-pre-wrap; /* Mozilla, since 1999 */
-        white-space: -pre-wrap; /* Opera 4-6 */
-        white-space: -o-pre-wrap; /* Opera 7 */
-        word-wrap: break-word; /* Internet Explorer 5.5+ */
     }
 </style>
